requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
defusedxml==0.7.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
pydantic==2.6.3
pydantic_core==2.16.3
pytest==8.1.1
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
vobject==0.9.6.1